requirements.txt
python-qpid-proton
influxdb